removing java webstart and cleaning up views
