+++ components/tools/OmeroWeb/omeroweb/webclient/templates/webclient/base/base.html
@@ -128,7 +128,6 @@
 
 
         <!-- Global Tools -->
-        {% include "webstart/includes/webstart_insight_toolbar.html" %}
         <ul class="header_toolbar" id="script_notifications">
                 {% include "webclient/base/includes/script_launch.html" %}
             {% include "webclient/base/includes/activities_info.html" %}
