+++ components/tools/OmeroWeb/omeroweb/webclient/templates/webclient/base/base_container.html
@@ -169,7 +169,6 @@
 {% block middle_header_right %}
 
         <!-- Global Tools -->
-        {% include "webstart/includes/webstart_insight_toolbar.html" %}
         <ul class="header_toolbar" id="script_notifications">
                 {% include "webclient/base/includes/script_launch.html" %}
             {% include "webclient/base/includes/activities_info.html" %}
