+++ components/tools/OmeroWeb/omeroweb/webclient/webclient_gateway.py
@@ -1780,121 +1780,7 @@ class OmeroWebGateway(omero.gateway.BlitzGateway):
         return share_id
 
     ##############################################
-    # History methods                        ##
-
-    # def getLastAcquiredImages (self):
-    #    tm = self.getTimelineService()
-    #    p = omero.sys.Parameters()
-    #    p.map = {}
-    #    f = omero.sys.Filter()
-    #    f.ownerId = rlong(self.getEventContext().userId)
-    #    f.groupId = rlong(self.getEventContext().groupId)
-    #    f.limit = rint(6)
-    #    p.theFilter = f
-    #    for e in tm.getMostRecentObjects(['Image'], p, False)["Image"]:
-    #        yield ImageWrapper(self, e)
-
-    def listLastImportedImages(self):
-        """
-        Retrieve most recent imported images
-        controlled by the security system.
-
-        @return:            Generator yielding Images
-        @rtype:             L{ImageWrapper} generator
-        """
-
-        tm = self.getTimelineService()
-        p = omero.sys.Parameters()
-        p.map = {}
-        f = omero.sys.Filter()
-        f.ownerId = rlong(self.getEventContext().userId)
-        f.groupId = rlong(self.getEventContext().groupId)
-        f.limit = rint(10)
-        p.theFilter = f
-        for e in tm.getMostRecentObjects(
-                ['Image'], p, False, self.SERVICE_OPTS)["Image"]:
-            yield ImageWrapper(self, e)
-
-    def listMostRecentShares(self):
-        """
-        Retrieve most recent shares
-        controlled by the security system.
-
-        @return:    Generator yielding SessionAnnotationLink
-        @rtype:     L{ShareWrapper} generator
-        """
-
-        tm = self.getTimelineService()
-        p = omero.sys.Parameters()
-        p.map = {}
-        f = omero.sys.Filter()
-        f.ownerId = rlong(self.getEventContext().userId)
-        f.limit = rint(10)
-        p.theFilter = f
-        for e in tm.getMostRecentShareCommentLinks(p, self.SERVICE_OPTS):
-            yield ShareWrapper(self, e.parent)
-
-    def listMostRecentShareComments(self):
-        """
-        Retrieve most recent share comments
-        controlled by the security system.
-
-        @return:    Generator yielding SessionAnnotationLink
-        @rtype:     L{SessionCommentWrapper} generator
-        """
-
-        tm = self.getTimelineService()
-        p = omero.sys.Parameters()
-        p.map = {}
-        f = omero.sys.Filter()
-        f.ownerId = rlong(self.getEventContext().userId)
-        f.limit = rint(10)
-        p.theFilter = f
-        for e in tm.getMostRecentShareCommentLinks(p, self.SERVICE_OPTS):
-            yield AnnotationWrapper(
-                self, e.child, link=ShareWrapper(self, e.parent))
-
-    def listMostRecentComments(self):
-        """
-        Retrieve most recent comment annotations
-        controlled by the security system.
-
-        @return:    Generator yielding BlitzObjectWrapper
-        @rtype:     L{BlitzObjectWrapper} generator
-        """
-
-        tm = self.getTimelineService()
-        p = omero.sys.Parameters()
-        p.map = {}
-        f = omero.sys.Filter()
-        f.ownerId = rlong(self.getEventContext().userId)
-        f.groupId = rlong(self.getEventContext().groupId)
-        f.limit = rint(10)
-        p.theFilter = f
-        for e in tm.getMostRecentAnnotationLinks(
-                None, ['CommentAnnotation'], None, p, self.SERVICE_OPTS):
-            yield omero.gateway.BlitzObjectWrapper(self, e)
-
-    def listMostRecentTags(self):
-        """
-        Retrieve most recent tag annotations
-        controlled by the security system.
-
-        @return:    Generator yielding BlitzObjectWrapper
-        @rtype:     L{BlitzObjectWrapper} generator
-        """
-
-        tm = self.getTimelineService()
-        p = omero.sys.Parameters()
-        p.map = {}
-        f = omero.sys.Filter()
-        # f.ownerId = rlong(self.getEventContext().userId)
-        f.groupId = rlong(self.getEventContext().groupId)
-        f.limit = rint(200)
-        p.theFilter = f
-        for e in tm.getMostRecentAnnotationLinks(
-                None, ['TagAnnotation'], None, p, self.SERVICE_OPTS):
-            yield omero.gateway.BlitzObjectWrapper(self, e.child)
+    # History                                   ##
 
     def getDataByPeriod(self, start, end, eid, otype=None, page=None):
         """
