--- components/tools/OmeroWeb/omeroweb/webstart/templates/webstart/includes/webstart_insight_script.html
@@ -1,27 +1,0 @@
-{% load i18n %}
-
-{% comment %}
-<!--
-  Copyright (C) 2011 University of Dundee & Open Microscopy Environment.
-  All rights reserved.
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU Affero General Public License as
-  published by the Free Software Foundation, either version 3 of the
-  License, or (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU Affero General Public License for more details.
-
-  You should have received a copy of the GNU Affero General Public License
-  along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-{% endcomment %}
-
-{% if insight_url %}
-    <a href="{{ insight_url }}"><img src="{% static "webstart/img/icon-omero-insight.png" %}"/></a>
-{% else %}
-    <h1 class="error">Webstart is not available</h1>
-{% endif %}
