--- components/tools/OmeroWeb/omeroweb/webstart/templates/webstart/includes/webstart_insight_toolbar.html
@@ -1,42 +1,0 @@
-{% load i18n %}
-
-{% comment %}
-<!--
-  Copyright (C) 2011-2013 University of Dundee & Open Microscopy Environment.
-  All rights reserved.
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU Affero General Public License as
-  published by the Free Software Foundation, either version 3 of the
-  License, or (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU Affero General Public License for more details.
-
-  You should have received a copy of the GNU Affero General Public License
-  along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-{% endcomment %}
-
-{% if insight_url %}
-<script>
-  var openWebstartPopup = function() {
-    // IE8 doesn't support arbitrary text for 'name' 2nd arg.  #6118
-    var owindow = window.open("{% url 'webstart_index' %}", '', 'height=500,width=500,left=550,top=50,toolbar=no,menubar=no,scrollbars=no,resizable=no,location=no,directories=no,status=no');
-    if(!owindow.closed) {
-        owindow.focus();
-    }
-    return false;
-};
-</script>
-
-<div class="header_toolbar" id="insight_button" title="Launch OMERO.insight via Java Webstart">
-
-  <a href="#" onClick="return openWebstartPopup()">
-    <img border="0" style=" height: 22px; padding: 1px 1px 0" src="{% static "webstart/img/icon-omero-insight-toolbar.png" %}">
-  </a>
-
-</div>
-{% endif %}
