--- components/tools/OmeroWeb/omeroweb/webstart/templates/webstart/index.html
@@ -1,61 +1,0 @@
-{% extends "webclient/login.html" %}
-{% load i18n %}
-
-{% comment %}
-<!--
-  Copyright (C) 2011-2013 University of Dundee & Open Microscopy Environment.
-  All rights reserved.
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU Affero General Public License as
-  published by the Free Software Foundation, either version 3 of the
-  License, or (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU Affero General Public License for more details.
-
-  You should have received a copy of the GNU Affero General Public License
-  along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-{% endcomment %}
-
-{% block title %}
-  OMERO webstart
-{% endblock %}
-
-{% block login %}
-<script>
-  $(function(){
-    if (window.opener && $.contains(window.opener.location, "webclient")) {
-        $("#goWebclientMsg span").text("return");
-        $("#goWebclientMsg a").click(function(){
-          window.close();
-          return false;
-        });
-    }
-  });
-</script>
-
-    {% if error %}<h1 class="error">Following error occured: WEBSTART_TEMPLATE, '{{ error }}'</h1>{% endif %}
-
-    <form class="standard_form inlined">
-      <h1 style="text-align: center">{% trans "Launch OMERO.insight via webstart" %}: </h1>
-      <p style="text-align: center">
-        {% include "webstart/includes/webstart_insight_script.html" %}
-      </p>
-      <p id="goWebclientMsg" style="text-align: center; margin-top: 15px; margin-bottom:0px; position: relative; top:15px">
-        To <span>go</span> to the webclient, click <a href="{% url 'webindex' %}">here</a>
-      </p>
-    </form>
-
-{% endblock %}
-
-
-
-
-
-
-
-
