--- components/tools/OmeroWeb/omeroweb/webstart/templates/webstart/insight.xml
@@ -1,40 +1,0 @@
-{% comment %}
-<!--
-  Copyright (C) 2011-2015 University of Dundee & Open Microscopy Environment.
-  All rights reserved.
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU Affero General Public License as
-  published by the Free Software Foundation, either version 3 of the
-  License, or (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU Affero General Public License for more details.
-
-  You should have received a copy of the GNU Affero General Public License
-  along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-{% endcomment %}<?xml version="1.0" encoding="utf-8"?>
-<jnlp codebase="{{ codebase }}" href="{{ href }}">
-  <information>
-    <title>{{title}}</title>
-    <vendor>{{vendor}}</vendor>
-    <homepage href="{{homepage}}"/>
-    <icon href="{% static icon %}"/>
-    <offline-allowed/>
-  </information>
-  <security>
-      <all-permissions/>
-  </security>
-  <resources>
-    <j2se version="1.6+" max-heap-size="{{heap}}"/>{% for jar in jarlist %}
-    <jar href="{{jar}}"/>{% endfor %}
-    <property name="jnlp.omero.host" value="{{host}}"/>
-    <property name="jnlp.omero.port" value="{{port}}"/>
-    <property name="jnlp.omero.web.host" value="{{web_host}}"/>{% if sessionid %}
-    <property name="jnlp.omero.sessionid" value="{{sessionid}}"/>{% endif %}
-  </resources>
-  <application-desc main-class="{{class}}"/>
-</jnlp>
