--- components/tools/OmeroWeb/omeroweb/webstart/urls.py
@@ -1,33 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-#
-#
-#
-# Copyright (c) 2008-2014 University of Dundee.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-# Author: Aleksandra Tarkowska <A(dot)Tarkowska(at)dundee(dot)ac(dot)uk>, 2008.
-#
-# Version: 1.0
-#
-
-from django.conf.urls import url, patterns
-from omeroweb.webstart import views
-
-urlpatterns = patterns(
-    'django.views.generic.simple',
-    url(r'^$', views.index, name="webstart_index"),
-    url(r'^jars/insight\.jnlp$', views.insight, name='webstart_insight'),
-)
